+++ tests/test_read.c
@@ -74,7 +74,7 @@ int test(LIBSSH2_SESSION *session)
         char buf[1024];
         ssize_t err = libssh2_channel_read(channel, buf, sizeof(buf));
         if(err < 0)
-            fprintf(stderr, "Unable to read response: %zd\n", err);
+            fprintf(stderr, "Unable to read response: %d\n", (int)err);
         else {
             int i;
             for(i = 0; i < err; ++i) {
